Sep 24 checkpoint has mutually consistent versions of MB and PB layouts, and pinouts documentation
completed layout of peripheral board, with a few retrofits to MB for IBC connector changes. Added a 4 position DIP switch to PB as a way to have jumpers that cross other traces, as I couldn't quite do a single sided layout. All switches go to the right when active, and the voltage regulators, whose 4 legs are represented by two 5 pin headers, are correctly scaled. I'd like to space out the top connector for the stepper motor connections, but the board is at the max size allowed, and Eagle CAD won't allow it to be moved upward.