PLU board release 2D
The release for PLU board is a set of 4 synchronized files.  Filenames are plu-rel2d, and extensions are brd, sch, dwf, and png. The png one is a Irfanview display of the DXF files illustraing that there are now pads for all components pins. This was achieved by editing brd and sch files (identically) to have explicit pad shapes for all pins, and not depend on handling of default case where pad shape isn't specified. Used "octagon" shape everywhere.